move util methods to Backend
